@@ -1,13 +1,14 @@
 const React = require('react')
+const PropTypes = require('prop-types')
 
 const RaisedButton = require('material-ui/RaisedButton').default
 
 class ShowMore extends React.Component {
   static get propTypes () {
     return {
-      defaultExpanded: React.PropTypes.bool,
-      hideLabel: React.PropTypes.string,
-      showLabel: React.PropTypes.string
+      defaultExpanded: PropTypes.bool,
+      hideLabel: PropTypes.string,
+      showLabel: PropTypes.string
     }
   }
 
@@ -44,7 +45,8 @@ class ShowMore extends React.Component {
         <RaisedButton
           className='control'
           onClick={this.handleClick}
-          label={label} />
+          label={label}
+        />
       </div>
     )
   }
