@@ -1,5 +1,5 @@
-const path = require('path')
 const React = require('react')
+const PropTypes = require('prop-types')
 
 const colors = require('material-ui/styles/colors')
 const Checkbox = require('material-ui/Checkbox').default
@@ -25,6 +25,9 @@ class PreferencesPage extends React.Component {
 
     this.handleStartupChange =
       this.handleStartupChange.bind(this)
+
+    this.handleSoundNotificationsChange =
+      this.handleSoundNotificationsChange.bind(this)
   }
 
   downloadPathSelector () {
@@ -33,11 +36,12 @@ class PreferencesPage extends React.Component {
         <PathSelector
           dialog={{
             title: 'Select download directory',
-            properties: [ 'openDirectory' ]
+            properties: ['openDirectory']
           }}
           onChange={this.handleDownloadPathChange}
           title='Download location'
-          value={this.props.state.saved.prefs.downloadPath} />
+          value={this.props.state.saved.prefs.downloadPath}
+        />
       </Preference>
     )
   }
@@ -52,8 +56,9 @@ class PreferencesPage extends React.Component {
         <Checkbox
           className='control'
           checked={!this.props.state.saved.prefs.openExternalPlayer}
-          label={'Play torrent media files using WebTorrent'}
-          onCheck={this.handleOpenExternalPlayerChange} />
+          label='Play torrent media files using WebTorrent'
+          onCheck={this.handleOpenExternalPlayerChange}
+        />
       </Preference>
     )
   }
@@ -68,7 +73,7 @@ class PreferencesPage extends React.Component {
         <Checkbox
           className='control'
           checked={this.props.state.saved.prefs.highestPlaybackPriority}
-          label={'Highest Playback Priority'}
+          label='Highest Playback Priority'
           onCheck={this.handleHighestPlaybackPriorityChange}
         />
         <p>Pauses all active torrents to allow playback to use all of the available bandwidth.</p>
@@ -94,12 +99,12 @@ class PreferencesPage extends React.Component {
         <PathSelector
           dialog={{
             title: 'Select media player app',
-            properties: [ 'openFile' ]
+            properties: ['openFile']
           }}
-          displayValue={playerName}
           onChange={this.handleExternalPlayerPathChange}
           title='External player'
-          value={playerPath ? path.dirname(playerPath) : null} />
+          value={playerPath}
+        />
       </Preference>
     )
   }
@@ -114,7 +119,7 @@ class PreferencesPage extends React.Component {
         <Checkbox
           className='control'
           checked={this.props.state.saved.prefs.autoAddTorrents}
-          label={'Watch for new .torrent files and add them immediately'}
+          label='Watch for new .torrent files and add them immediately'
           onCheck={(e, value) => { this.handleAutoAddTorrentsChange(e, value) }}
         />
       </Preference>
@@ -132,11 +137,11 @@ class PreferencesPage extends React.Component {
     dispatch('updatePreferences', 'autoAddTorrents', isChecked)
 
     if (isChecked) {
-      dispatch('startFolderWatcher', null)
+      dispatch('startFolderWatcher')
       return
     }
 
-    dispatch('stopFolderWatcher', null)
+    dispatch('stopFolderWatcher')
   }
 
   torrentsFolderPathSelector () {
@@ -147,17 +152,17 @@ class PreferencesPage extends React.Component {
         <PathSelector
           dialog={{
             title: 'Select folder to watch for new torrents',
-            properties: [ 'openDirectory' ]
+            properties: ['openDirectory']
           }}
-          displayValue={torrentsFolderPath || ''}
-          onChange={this.handletorrentsFolderPathChange}
+          onChange={this.handleTorrentsFolderPathChange}
           title='Folder to watch'
-          value={torrentsFolderPath ? path.dirname(torrentsFolderPath) : null} />
+          value={torrentsFolderPath}
+        />
       </Preference>
     )
   }
 
-  handletorrentsFolderPathChange (filePath) {
+  handleTorrentsFolderPathChange (filePath) {
     dispatch('updatePreferences', 'torrentsFolderPath', filePath)
   }
 
@@ -176,7 +181,8 @@ class PreferencesPage extends React.Component {
         <RaisedButton
           className='control'
           onClick={this.handleSetDefaultApp}
-          label='Make WebTorrent the default' />
+          label='Make WebTorrent the default'
+        />
       </Preference>
     )
   }
@@ -185,23 +191,38 @@ class PreferencesPage extends React.Component {
     dispatch('updatePreferences', 'startup', isChecked)
   }
 
-  setStartupSection () {
+  setStartupCheckbox () {
     if (config.IS_PORTABLE) {
       return
     }
 
     return (
-      <PreferencesSection title='Startup'>
-        <Preference>
-          <Checkbox
-            className='control'
-            checked={this.props.state.saved.prefs.startup}
-            label={'Open WebTorrent on startup.'}
-            onCheck={this.handleStartupChange}
-          />
-        </Preference>
-      </PreferencesSection>
+      <Preference>
+        <Checkbox
+          className='control'
+          checked={this.props.state.saved.prefs.startup}
+          label='Open WebTorrent on startup'
+          onCheck={this.handleStartupChange}
+        />
+      </Preference>
     )
+  }
+
+  soundNotificationsCheckbox () {
+    return (
+      <Preference>
+        <Checkbox
+          className='control'
+          checked={this.props.state.saved.prefs.soundNotifications}
+          label='Enable sounds'
+          onCheck={this.handleSoundNotificationsChange}
+        />
+      </Preference>
+    )
+  }
+
+  handleSoundNotificationsChange (e, isChecked) {
+    dispatch('updatePreferences', 'soundNotifications', isChecked)
   }
 
   handleSetDefaultApp () {
@@ -229,7 +250,10 @@ class PreferencesPage extends React.Component {
         <PreferencesSection title='Default torrent app'>
           {this.setDefaultAppButton()}
         </PreferencesSection>
-        {this.setStartupSection()}
+        <PreferencesSection title='General'>
+          {this.setStartupCheckbox()}
+          {this.soundNotificationsCheckbox()}
+        </PreferencesSection>
       </div>
     )
   }
@@ -238,7 +262,7 @@ class PreferencesPage extends React.Component {
 class PreferencesSection extends React.Component {
   static get propTypes () {
     return {
-      title: React.PropTypes.string
+      title: PropTypes.string
     }
   }
 
