@@ -77,8 +77,8 @@ function onPlayerPlay () {
 
 function onPlayerUpdate (state) {
   if (!isEnabled()) return
-  buttons[PREV].flags = [ state.hasPrevious ? 'enabled' : 'disabled' ]
-  buttons[NEXT].flags = [ state.hasNext ? 'enabled' : 'disabled' ]
+  buttons[PREV].flags = [state.hasPrevious ? 'enabled' : 'disabled']
+  buttons[NEXT].flags = [state.hasNext ? 'enabled' : 'disabled']
   update()
 }
 
