@@ -13,14 +13,13 @@ module.exports = class SubtitlesController {
   }
 
   openSubtitles () {
-    remote.dialog.showOpenDialog({
+    const filenames = remote.dialog.showOpenDialogSync({
       title: 'Select a subtitles file.',
-      filters: [ { name: 'Subtitles', extensions: ['vtt', 'srt'] } ],
-      properties: [ 'openFile' ]
-    }, (filenames) => {
-      if (!Array.isArray(filenames)) return
-      this.addSubtitles(filenames, true)
+      filters: [{ name: 'Subtitles', extensions: ['vtt', 'srt'] }],
+      properties: ['openFile']
     })
+    if (!Array.isArray(filenames)) return
+    this.addSubtitles(filenames, true)
   }
 
   selectSubtitle (ix) {
