@@ -38,7 +38,8 @@ function getPreviousIndex (state) {
 
 function getCurrentLocalURL (state) {
   return state.server
-    ? state.server.localURL + '/' + state.playing.fileIndex
+    ? state.server.localURL + '/' + state.playing.fileIndex + '/' +
+      encodeURIComponent(state.playing.fileName)
     : ''
 }
 
