@@ -1,7 +1,6 @@
 const React = require('react')
 const Bitfield = require('bitfield')
 const prettyBytes = require('prettier-bytes')
-const zeroFill = require('zero-fill')
 
 const TorrentSummary = require('../lib/torrent-summary')
 const Playlist = require('../lib/playlist')
@@ -20,7 +19,8 @@ module.exports = class Player extends React.Component {
       <div
         className='player'
         onWheel={handleVolumeWheel}
-        onMouseMove={dispatcher('mediaMouseMoved')}>
+        onMouseMove={dispatcher('mediaMouseMoved')}
+      >
         {showVideo ? renderMedia(state) : renderCastScreen(state)}
         {showControls ? renderPlayerControls(state) : null}
       </div>
@@ -95,6 +95,13 @@ function renderMedia (state) {
       delete file.selectedSubtitle
     }
 
+    // Switch to selected audio track
+    const audioTracks = mediaElement.audioTracks || []
+    for (let j = 0; j < audioTracks.length; j++) {
+      const isSelectedTrack = j === state.playing.audioTracks.selectedIndex
+      audioTracks[j].enabled = isSelectedTrack
+    }
+
     state.playing.volume = mediaElement.volume
   }
 
@@ -107,10 +114,11 @@ function renderMedia (state) {
       trackTags.push(
         <track
           key={i}
-          default={isSelected ? 'default' : ''}
+          default={isSelected}
           label={track.label}
-          type='subtitles'
-          src={track.buffer} />
+          kind='subtitles'
+          src={track.buffer}
+        />
       )
     }
   }
@@ -127,7 +135,7 @@ function renderMedia (state) {
       onError={dispatcher('mediaError')}
       onTimeUpdate={dispatcher('mediaTimeUpdate')}
       onEncrypted={dispatcher('mediaEncrypted')}
-      onCanPlay={onCanPlay}>
+    >
       {trackTags}
     </MediaTagName>
   )
@@ -137,21 +145,57 @@ function renderMedia (state) {
     <div
       key='letterbox'
       className='letterbox'
-      onMouseMove={dispatcher('mediaMouseMoved')}>
+      onMouseMove={dispatcher('mediaMouseMoved')}
+    >
       {mediaTag}
       {renderOverlay(state)}
     </div>
   )
 
-  // As soon as we know the video dimensions, resize the window
   function onLoadedMetadata (e) {
-    if (state.playing.type !== 'video') return
-    const video = e.target
-    const dimensions = {
-      width: video.videoWidth,
-      height: video.videoHeight
+    const mediaElement = e.target
+
+    // check if we can decode video and audio track
+    if (state.playing.type === 'video') {
+      if (mediaElement.videoTracks.length === 0) {
+        dispatch('mediaError', 'Video codec unsupported')
+      }
+
+      if (mediaElement.audioTracks.length === 0) {
+        dispatch('mediaError', 'Audio codec unsupported')
+      }
+
+      dispatch('mediaSuccess')
+
+      const dimensions = {
+        width: mediaElement.videoWidth,
+        height: mediaElement.videoHeight
+      }
+
+      // As soon as we know the video dimensions, resize the window
+      dispatch('setDimensions', dimensions)
+
+      // set audioTracks
+      const tracks = []
+      for (let i = 0; i < mediaElement.audioTracks.length; i++) {
+        tracks.push({
+          label: mediaElement.audioTracks[i].label || `Track ${i + 1}`,
+          language: mediaElement.audioTracks[i].language
+        })
+      }
+
+      state.playing.audioTracks.tracks = tracks
+      state.playing.audioTracks.selectedIndex = 0
     }
-    dispatch('setDimensions', dimensions)
+
+    // check if we can decode audio track
+    if (state.playing.type === 'audio') {
+      if (mediaElement.audioTracks.length === 0) {
+        dispatch('mediaError', 'Audio codec unsupported')
+      }
+
+      dispatch('mediaSuccess')
+    }
   }
 
   function onEnded (e) {
@@ -161,19 +205,6 @@ function renderMedia (state) {
       // When the last video completes, pause the video instead of looping
       state.playing.isPaused = true
       if (state.window.isFullScreen) dispatch('toggleFullScreen')
-    }
-  }
-
-  function onCanPlay (e) {
-    const elem = e.target
-    if (state.playing.type === 'video' &&
-      elem.webkitVideoDecodedByteCount === 0) {
-      dispatch('mediaError', 'Video codec unsupported')
-    } else if (elem.webkitAudioDecodedByteCount === 0) {
-      dispatch('mediaError', 'Audio codec unsupported')
-    } else {
-      dispatch('mediaSuccess')
-      elem.play()
     }
   }
 }
@@ -293,7 +324,7 @@ function renderAudioMetadata (state) {
     }
     elems.push((
       <div key='release' className='audio-release'>
-        <label>Release</label>{ releaseInfo.join(', ') }
+        <label>Release</label>{releaseInfo.join(', ')}
       </div>
     ))
   }
@@ -301,14 +332,18 @@ function renderAudioMetadata (state) {
   // Audio metadata: format
   const format = []
   fileSummary.audioInfo.format = fileSummary.audioInfo.format || ''
-  if (fileSummary.audioInfo.format.dataformat) {
-    format.push(fileSummary.audioInfo.format.dataformat)
+  if (fileSummary.audioInfo.format.container) {
+    format.push(fileSummary.audioInfo.format.container)
+  }
+  if (fileSummary.audioInfo.format.codec &&
+    fileSummary.audioInfo.format.container !== fileSummary.audioInfo.format.codec) {
+    format.push(fileSummary.audioInfo.format.codec)
   }
   if (fileSummary.audioInfo.format.bitrate) {
     format.push(Math.round(fileSummary.audioInfo.format.bitrate / 1000) + ' kbps') // 128 kbps
   }
   if (fileSummary.audioInfo.format.sampleRate) {
-    format.push(Math.round(fileSummary.audioInfo.format.sampleRate / 100) / 10 + ' kHz') // 44.1 kHz
+    format.push(Math.round(fileSummary.audioInfo.format.sampleRate / 100) / 10 + ' kHz')
   }
   if (fileSummary.audioInfo.format.bitsPerSample) {
     format.push(fileSummary.audioInfo.format.bitsPerSample + ' bit')
@@ -316,7 +351,7 @@ function renderAudioMetadata (state) {
   if (format.length > 0) {
     elems.push((
       <div key='format' className='audio-format'>
-        <label>Format</label>{ format.join(', ') }
+        <label>Format</label>{format.join(', ')}
       </div>
     ))
   }
@@ -358,7 +393,7 @@ function renderLoadingSpinner (state) {
     <div key='loading' className='media-stalled'>
       <div key='loading-spinner' className='loading-spinner' />
       <div key='loading-progress' className='loading-status ellipsis'>
-        <span className='progress'>{fileProgress}%</span> downloaded
+        <span><span className='progress'>{fileProgress}%</span> downloaded</span>
         <span> ↓ {prettyBytes(prog.downloadSpeed || 0)}/s</span>
         <span> ↑ {prettyBytes(prog.uploadSpeed || 0)}/s</span>
       </div>
@@ -425,6 +460,7 @@ function renderCastOptions (state) {
     return (
       <li key={ix} onClick={dispatcher('selectCastDevice', ix)}>
         <i className='icon'>{isSelected ? 'radio_button_checked' : 'radio_button_unchecked'}</i>
+        {' '}
         {name}
       </li>
     )
@@ -464,6 +500,27 @@ function renderSubtitleOptions (state) {
   )
 }
 
+function renderAudioTrackOptions (state) {
+  const audioTracks = state.playing.audioTracks
+  if (!audioTracks.tracks.length || !audioTracks.showMenu) return
+
+  const items = audioTracks.tracks.map(function (track, ix) {
+    const isSelected = state.playing.audioTracks.selectedIndex === ix
+    return (
+      <li key={ix} onClick={dispatcher('selectAudioTrack', ix)}>
+        <i className='icon'>{'radio_button_' + (isSelected ? 'checked' : 'unchecked')}</i>
+        {track.label}
+      </li>
+    )
+  })
+
+  return (
+    <ul key='audio-track-options' className='options-list'>
+      {items}
+    </ul>
+  )
+}
+
 function renderPlayerControls (state) {
   const positionPercent = 100 * state.playing.currentTime / state.playing.duration
   const playbackCursorStyle = { left: 'calc(' + positionPercent + '% - 3px)' }
@@ -472,6 +529,9 @@ function renderPlayerControls (state) {
     : state.playing.subtitles.selectedIndex >= 0
       ? 'active'
       : ''
+  const multiAudioClass = state.playing.audioTracks.tracks.length > 1
+    ? 'active'
+    : 'disabled'
   const prevClass = Playlist.hasPrevious(state) ? '' : 'disabled'
   const nextClass = Playlist.hasNext(state) ? '' : 'disabled'
 
@@ -481,41 +541,47 @@ function renderPlayerControls (state) {
       <div
         key='cursor'
         className='playback-cursor'
-        style={playbackCursorStyle} />
+        style={playbackCursorStyle}
+      />
       <div
         key='scrub-bar'
         className='scrub-bar'
-        draggable='true'
+        draggable
         onDragStart={handleDragStart}
         onClick={handleScrub}
-        onDrag={handleScrub} />
+        onDrag={handleScrub}
+      />
     </div>,
 
     <i
       key='skip-previous'
       className={'icon skip-previous float-left ' + prevClass}
-      onClick={dispatcher('previousTrack')}>
+      onClick={dispatcher('previousTrack')}
+    >
       skip_previous
     </i>,
 
     <i
       key='play'
       className='icon play-pause float-left'
-      onClick={dispatcher('playPause')}>
+      onClick={dispatcher('playPause')}
+    >
       {state.playing.isPaused ? 'play_arrow' : 'pause'}
     </i>,
 
     <i
       key='skip-next'
       className={'icon skip-next float-left ' + nextClass}
-      onClick={dispatcher('nextTrack')}>
+      onClick={dispatcher('nextTrack')}
+    >
       skip_next
     </i>,
 
     <i
       key='fullscreen'
       className='icon fullscreen float-right'
-      onClick={dispatcher('toggleFullScreen')}>
+      onClick={dispatcher('toggleFullScreen')}
+    >
       {state.window.isFullScreen ? 'fullscreen_exit' : 'fullscreen'}
     </i>
   ]
@@ -526,8 +592,17 @@ function renderPlayerControls (state) {
       <i
         key='subtitles'
         className={'icon closed-caption float-right ' + captionsClass}
-        onClick={handleSubtitles}>
+        onClick={handleSubtitles}
+      >
         closed_caption
+      </i>
+    ), (
+      <i
+        key='audio-tracks'
+        className={'icon multi-audio float-right ' + multiAudioClass}
+        onClick={handleAudioTracks}
+      >
+        library_music
       </i>
     ))
   }
@@ -539,9 +614,9 @@ function renderPlayerControls (state) {
 
   // Add the cast buttons. Icons for each cast type, connected/disconnected:
   const buttonIcons = {
-    'chromecast': { true: 'cast_connected', false: 'cast' },
-    'airplay': { true: 'airplay', false: 'airplay' },
-    'dlna': { true: 'tv', false: 'tv' }
+    chromecast: { true: 'cast_connected', false: 'cast' },
+    airplay: { true: 'airplay', false: 'airplay' },
+    dlna: { true: 'tv', false: 'tv' }
   }
   castTypes.forEach(function (castType) {
     // Do we show this button (eg. the Chromecast button) at all?
@@ -570,7 +645,8 @@ function renderPlayerControls (state) {
       <i
         key={castType}
         className={'icon device float-right ' + buttonClass}
-        onClick={buttonHandler}>
+        onClick={buttonHandler}
+      >
         {buttonIcon}
       </i>
     ))
@@ -593,7 +669,8 @@ function renderPlayerControls (state) {
     <div key='volume' className='volume float-left'>
       <i
         className='icon volume-icon float-left'
-        onMouseDown={handleVolumeMute}>
+        onMouseDown={handleVolumeMute}
+      >
         {volumeIcon}
       </i>
       <input
@@ -601,7 +678,8 @@ function renderPlayerControls (state) {
         type='range' min='0' max='1' step='0.05'
         value={volume}
         onChange={handleVolumeScrub}
-        style={volumeStyle} />
+        style={volumeStyle}
+      />
     </div>
   ))
 
@@ -624,12 +702,15 @@ function renderPlayerControls (state) {
   }
 
   return (
-    <div key='controls' className='controls'
+    <div
+      key='controls' className='controls'
       onMouseEnter={dispatcher('mediaControlsMouseEnter')}
-      onMouseLeave={dispatcher('mediaControlsMouseLeave')}>
+      onMouseLeave={dispatcher('mediaControlsMouseLeave')}
+    >
       {elements}
       {renderCastOptions(state)}
       {renderSubtitleOptions(state)}
+      {renderAudioTrackOptions(state)}
     </div>
   )
 
@@ -672,6 +753,10 @@ function renderPlayerControls (state) {
     } else {
       dispatch('toggleSubtitlesMenu')
     }
+  }
+
+  function handleAudioTracks (e) {
+    dispatch('toggleAudioTracksMenu')
   }
 }
 
@@ -734,14 +819,14 @@ function formatTime (time, total) {
     return '0:00'
   }
 
-  let totalHours = Math.floor(total / 3600)
-  let totalMinutes = Math.floor(total % 3600 / 60)
-  let hours = Math.floor(time / 3600)
+  const totalHours = Math.floor(total / 3600)
+  const totalMinutes = Math.floor(total / 60)
+  const hours = Math.floor(time / 3600)
   let minutes = Math.floor(time % 3600 / 60)
-  if (totalMinutes > 9) {
-    minutes = zeroFill(2, minutes)
+  if (totalMinutes > 9 && minutes < 10) {
+    minutes = '0' + minutes
   }
-  let seconds = zeroFill(2, Math.floor(time % 60))
+  const seconds = `0${Math.floor(time % 60)}`.slice(-2)
 
   return (totalHours > 0 ? hours + ':' : '') + minutes + ':' + seconds
 }
