@@ -1,15 +1,18 @@
 const electron = require('electron')
 const config = require('./config')
 
-console.log('Mocking electron.dialog.showOpenDialog...')
-electron.dialog.showOpenDialog = function (win, opts, cb) {
-  const ret = /select.*torrent file/i.test(opts.title)
+console.log('Mocking electron.dialog.showOpenDialogSync...')
+electron.dialog.showOpenDialogSync = showOpenDialogSync
+
+function showOpenDialogSync (win, opts) {
+  return /select.*torrent file/i.test(opts.title)
     ? config.TORRENT_FILES
     : config.SEED_FILES
-  cb(ret)
 }
 
-console.log('Mocking electron.remote.dialog.showSaveDialog...')
-electron.dialog.showSaveDialog = function (win, opts, cb) {
-  cb(config.SAVED_TORRENT_FILE)
+console.log('Mocking electron.dialog.showSaveDialogSync...')
+electron.dialog.showSaveDialogSync = showSaveDialogSync
+
+function showSaveDialogSync (win, opts) {
+  return config.SAVED_TORRENT_FILE
 }
